Fix bug using long long double instead of long long

+++ 0x00-hello_world/6-size.c
@@ -37,7 +37,7 @@ void print_type(char *aan, char *type)
 	if (!strcmp(type, "long int"))
 		size = sizeof(long int);
 	if (!strcmp(type, "long long int"))
-		size = sizeof(long long double);
+		size = sizeof(long long);
 	if (!strcmp(type, "float"))
 		size = sizeof(float);
